The 2 demo services are ready.
Some string changes in the demo storage test file.
Fixed a bug in transfer delegate where a path was constructed with
File.PathSeparator (on linux ":") instead of File.separator (in Linux
"/")
The two services are working correctly.
The StorageTest searches a file that is at the localhost.
The TransferTest downloads a file from the localhost (from the shared
dir to the download dir)