Bugfix in RepairDelegate.pushSolution(...)
When the solution is ready to be pushed, each level of the routing table
was retrieved. There wasn't a check to determine if the level was empty.
The result was shown during a multiple peer failure where a peer should
push the solution of a subtree to an already failed peer.